Improve win32_edge_engine window behavior
* Window will respect Windows theme background color

This does not also apply to the webview control tool (that requires further work). So, window color flash on load is reduced but not elimited.

* Windows are centered centered on the screen.

Idealy they would also be able to accept an initial size. This is at least better than them rendering in a random position on screen.